requirements.txt
Flask
requests 
google-generativeai 
PyPDF2 
sentence-transformers 
chromadb 
mysql-connector-python 
APScheduler 